+++ src/app/coreFilesUpdater/BaseCoreFilesUpdater.ts
@@ -3,10 +3,11 @@
 // Licensed under the MIT license.
 
 import { GeneratorTeamsAppOptions } from "../GeneratorTeamsAppOptions";
+import { Editor } from 'mem-fs-editor';
 import * as Generator from 'yeoman-generator';
 
 export abstract class BaseCoreFilesUpdater {
-    public updateCoreFiles(options: GeneratorTeamsAppOptions, fs: Generator.MemFsEditor): boolean {
+    public updateCoreFiles(options: GeneratorTeamsAppOptions, fs: Editor): boolean {
         // nop
         return true;
     }