+++ src/app/coreFilesUpdater/coreFilesUpdaters/CoreFilesUpdaterNop.ts
@@ -5,11 +5,12 @@
 import { GeneratorTeamsAppOptions } from "../../GeneratorTeamsAppOptions";
 import { BaseCoreFilesUpdater } from "../BaseCoreFilesUpdater";
 import * as Generator from 'yeoman-generator';
+import { Editor } from 'mem-fs-editor';
 import { Project, PropertyAssignment } from 'ts-morph';
 import * as ts from 'typescript';
 
 export class CoreFilesUpdaterNop extends BaseCoreFilesUpdater {
-    public updateCoreFiles(options: GeneratorTeamsAppOptions, fs: Generator.MemFsEditor): boolean {
+    public updateCoreFiles(options: GeneratorTeamsAppOptions, fs: Editor): boolean {
         return true;
     }
 }