+++ tests/tabGenerator.spec.ts
@@ -1156,6 +1156,68 @@ describe("teams:tab", function () {
     }
   });
 
+  it("should generate tab project with SSO support (schema 1.7)", async () => {
+    const TABSSOAPPID = "00000000-0000-0000-0000-000000000123";
+    const TABSSOAPPURI = "api://tabtest01.azurewebsites.net/00000000-0000-0000-0000-000000000123";
+    const projectPath = testHelper.TEMP_TAB_GENERATOR_PATH + "/tab-17-SSO"
+    await helpers
+      .run(testHelper.GENERATOR_PATH)
+      .inDir(projectPath)
+      .withArguments(["--no-telemetry"])
+      .withPrompts({
+        solutionName: "tab-test-01-sso",
+        whichFolder: "current",
+        name: "tabtest01",
+        developer: "generator teams developer",
+        manifestVersion: "v1.7",
+        parts: "tab",
+        unitTestsEnabled: false,
+        tabSharePoint: false,
+        tabSSO: true,
+        tabSSOAppId: TABSSOAPPID,
+        tabSSOAppUri: TABSSOAPPURI,
+        tabType: "configurable"
+      })
+      .withGenerators(testHelper.DEPENDENCIES);
+
+    assert.file(testHelper.ROOT_FILES);
+    assert.noFile(testHelper.TEST_FILES);
+    assert.file(testHelper.APP_FILES);
+    assert.file(testHelper.SCRIPT_FILES);
+    assert.file(testHelper.WEB_FILES);
+    assert.file(testHelper.MANIFEST_FILES);
+
+    assert.fileContent("src/manifest/manifest.json", testHelper.SCHEMA_17);
+    assert.jsonFileContent("src/manifest/manifest.json", {
+      webApplicationInfo:
+      {
+        id: "{{TABTEST01TAB_APP_ID}}",
+        resource: "{{TABTEST01TAB_APP_URI}}"
+      }
+
+    });
+
+    assert.jsonFileContent("src/manifest/manifest.json", {
+      isFullScreen: false
+    });
+
+    assert.fileContent(".env", `TABTEST01TAB_APP_ID=${TABSSOAPPID}`);
+    assert.fileContent(".env", `TABTEST01TAB_APP_URI=${TABSSOAPPURI}`);
+
+    assert.file(TAB_HTML_FILES);
+    assert.file(TAB_FILES);
+    assert.file(TAB_SCRIPT_FILES);
+    assert.noFile(TAB_SCRIPT_TEST_FILES);
+
+    if (process.env.TEST_TYPE == testHelper.TestTypes.INTEGRATION) {
+      const npmInstallResult = await testHelper.runNpmCommand("npm install --prefer-offline", projectPath);
+      assert.equal(false, npmInstallResult);
+
+      const npmRunBuildResult = await testHelper.runNpmCommand("npm run build", projectPath);
+      assert.equal(false, npmRunBuildResult);
+    }
+  });
+
   it("should generate tab project with SSO support (schema devPreview)", async () => {
     const TABSSOAPPID = "00000000-0000-0000-0000-000000000123";
     const TABSSOAPPURI = "api://tabtest01.azurewebsites.net/00000000-0000-0000-0000-000000000123";
@@ -1320,6 +1382,114 @@ describe("teams:tab", function () {
     }
   });
 
+  it("should generate tab project with full screen set to true (schema 1.7)", async () => {
+    const TABSSOAPPID = "00000000-0000-0000-0000-000000000123";
+    const TABSSOAPPURI = "api://tabtest01.azurewebsites.net/00000000-0000-0000-0000-000000000123";
+    const projectPath = testHelper.TEMP_TAB_GENERATOR_PATH + "/tab-17-full-screen"
+    await helpers
+      .run(testHelper.GENERATOR_PATH)
+      .inDir(projectPath)
+      .withArguments(["--no-telemetry"])
+      .withPrompts({
+        solutionName: "tab-test-01-sso",
+        whichFolder: "current",
+        name: "tabtest01",
+        developer: "generator teams developer",
+        manifestVersion: "v1.7",
+        parts: "tab",
+        unitTestsEnabled: false,
+        tabSharePoint: false,
+        tabSSO: true,
+        tabSSOAppId: TABSSOAPPID,
+        tabSSOAppUri: TABSSOAPPURI,
+        tabType: "configurable",
+        isFullScreen: true
+      })
+      .withGenerators(testHelper.DEPENDENCIES);
+
+    assert.file(testHelper.ROOT_FILES);
+    assert.noFile(testHelper.TEST_FILES);
+    assert.file(testHelper.APP_FILES);
+    assert.file(testHelper.SCRIPT_FILES);
+    assert.file(testHelper.WEB_FILES);
+    assert.file(testHelper.MANIFEST_FILES);
+
+    assert.fileContent("src/manifest/manifest.json", testHelper.SCHEMA_17);
+    assert.jsonFileContent("src/manifest/manifest.json", {
+      isFullScreen: true
+    });
+    assert.fileContent(".env", `APPLICATION_ID=`);
+    assert.fileContent(".env", `PACKAGE_NAME=`);
+
+    assert.file(TAB_HTML_FILES);
+    assert.file(TAB_FILES);
+    assert.file(TAB_SCRIPT_FILES);
+    assert.noFile(TAB_SCRIPT_TEST_FILES);
+
+    if (process.env.TEST_TYPE == testHelper.TestTypes.INTEGRATION) {
+      const npmInstallResult = await testHelper.runNpmCommand("npm install --prefer-offline", projectPath);
+      assert.equal(false, npmInstallResult);
+
+      const npmRunBuildResult = await testHelper.runNpmCommand("npm run build", projectPath);
+      assert.equal(false, npmRunBuildResult);
+
+    }
+  });
+
+  it("should generate tab project with full screen set to false (schema 1.7)", async () => {
+    const TABSSOAPPID = "00000000-0000-0000-0000-000000000123";
+    const TABSSOAPPURI = "api://tabtest01.azurewebsites.net/00000000-0000-0000-0000-000000000123";
+    const projectPath = testHelper.TEMP_TAB_GENERATOR_PATH + "/tab-17-no-full-screen"
+    await helpers
+      .run(testHelper.GENERATOR_PATH)
+      .inDir(projectPath)
+      .withArguments(["--no-telemetry"])
+      .withPrompts({
+        solutionName: "tab-test-01-sso",
+        whichFolder: "current",
+        name: "tabtest01",
+        developer: "generator teams developer",
+        manifestVersion: "v1.7",
+        parts: "tab",
+        unitTestsEnabled: false,
+        tabSharePoint: false,
+        tabSSO: true,
+        tabSSOAppId: TABSSOAPPID,
+        tabSSOAppUri: TABSSOAPPURI,
+        tabType: "configurable",
+        isFullScreen: false
+      })
+      .withGenerators(testHelper.DEPENDENCIES);
+
+    assert.file(testHelper.ROOT_FILES);
+    assert.noFile(testHelper.TEST_FILES);
+    assert.file(testHelper.APP_FILES);
+    assert.file(testHelper.SCRIPT_FILES);
+    assert.file(testHelper.WEB_FILES);
+    assert.file(testHelper.MANIFEST_FILES);
+
+    assert.fileContent("src/manifest/manifest.json", testHelper.SCHEMA_17);
+    assert.jsonFileContent("src/manifest/manifest.json", {
+      isFullScreen: false
+    });
+    assert.fileContent(".env", `APPLICATION_ID=`);
+    assert.fileContent(".env", `PACKAGE_NAME=`);
+
+    assert.file(TAB_HTML_FILES);
+    assert.file(TAB_FILES);
+    assert.file(TAB_SCRIPT_FILES);
+    assert.noFile(TAB_SCRIPT_TEST_FILES);
+
+    if (process.env.TEST_TYPE == testHelper.TestTypes.INTEGRATION) {
+      const npmInstallResult = await testHelper.runNpmCommand("npm install --prefer-offline", projectPath);
+      assert.equal(false, npmInstallResult);
+
+      const npmRunBuildResult = await testHelper.runNpmCommand("npm run build", projectPath);
+      assert.equal(false, npmRunBuildResult);
+
+    }
+  });
+
   it("should generate tab project with schema 1.5 and upgrade to schema 1.6", async () => {
     const projectPath = testHelper.TEMP_TAB_GENERATOR_PATH + "/tab-15-to-16-upgrade";
     await helpers
@@ -1362,6 +1532,63 @@ describe("teams:tab", function () {
       .withGenerators(testHelper.DEPENDENCIES);
 
     assert.fileContent("src/manifest/manifest.json", testHelper.SCHEMA_16);
+
+    if (process.env.TEST_TYPE == testHelper.TestTypes.INTEGRATION) {
+      const npmInstallResult = await testHelper.runNpmCommand("npm install --prefer-offline", projectPath);
+      assert.equal(false, npmInstallResult);
+
+      const npmRunBuildResult = await testHelper.runNpmCommand("npm run build", projectPath);
+      assert.equal(false, npmRunBuildResult);
+    }
+
+  });
+
+  it("should generate tab project with schema 1.6 and upgrade to schema 1.7", async () => {
+    const projectPath = testHelper.TEMP_TAB_GENERATOR_PATH + "/tab-16-to-17-upgrade";
+    await helpers
+      .run(testHelper.GENERATOR_PATH)
+      .inDir(projectPath)
+      .withArguments(["--no-telemetry"])
+      .withPrompts({
+        solutionName: "tab-test-01-sso",
+        whichFolder: "current",
+        name: "tabtest01",
+        developer: "generator teams developer",
+        manifestVersion: "v1.6",
+        parts: "tab",
+        unitTestsEnabled: false,
+        tabSharePoint: false,
+        tabType: "configurable"
+      })
+      .withGenerators(testHelper.DEPENDENCIES);
+    assert.fileContent("src/manifest/manifest.json", testHelper.SCHEMA_16);
+
+    if (process.env.TEST_TYPE == testHelper.TestTypes.INTEGRATION) {
+      const npmInstallResult = await testHelper.runNpmCommand("npm install --prefer-offline", projectPath);
+      assert.equal(false, npmInstallResult);
+
+      const npmRunBuildResult = await testHelper.runNpmCommand("npm run build", projectPath);
+      assert.equal(false, npmRunBuildResult);
+    }
+
+    await helpers
+      .run(testHelper.GENERATOR_PATH)
+      .cd(projectPath)
+      .withArguments(["--no-telemetry"])
+      .withPrompts({
+        manifestVersion: "v1.7",
+        confirmedAdd: true,
+        updateBuildSystem: false,
+        updateManifestVersion: true,
+        parts: ""
+      })
+      .withGenerators(testHelper.DEPENDENCIES);
+
+    assert.fileContent("src/manifest/manifest.json", testHelper.SCHEMA_17);
+
+    assert.jsonFileContent("src/manifest/manifest.json", {
+      isFullScreen: undefined
+    });
 
     if (process.env.TEST_TYPE == testHelper.TestTypes.INTEGRATION) {
       const npmInstallResult = await testHelper.runNpmCommand("npm install --prefer-offline", projectPath);